llm_game_cli_windsurf/requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langgraph>=0.0.15
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langchain-anthropic>=0.0.5
typing-extensions>=4.8.0
python-dotenv>=1.0.0
